refactor: AuthorizationGuardWithContext を削除し AuthorizationGuard に統合する
AnswerEntrySet 集約の導入により、認可に必要な状態が集約内部に取り込まれ
外部 Context を注入する AuthorizationGuardWithContext が不要になった。
AuthorizationGuard を WithContext への委譲から直接実装に書き換え、
Action マーカー型 (Create/Read/Update/Delete) も同ファイルへ移設した。
