+++ server/domain/src/repository/form/active_form_repository.rs
@@ -6,7 +6,7 @@ use crate::{
     form::models::{ActiveForm, FormId},
     types::{
         authorization_guard::AuthorizationGuard,
-        authorization_guard_with_context::{Create, Read, Update},
+        authorization_guard::{Create, Read, Update},
     },
     user::models::ActiveUser,
 };
