+++ server/domain/src/repository/form/answer_entry_set_repository.rs
@@ -9,7 +9,7 @@ use crate::{
         comment::models::{Comment, CommentId},
     },
     types::authorization_guard::AuthorizationGuard,
-    types::authorization_guard_with_context::{Create, Read, Update},
+    types::authorization_guard::{Create, Read, Update},
     user::models::Actor,
 };
