+++ server/domain/src/repository/notification_repository.rs
@@ -7,7 +7,7 @@ use crate::{
     notification::models::NotificationPreference,
     types::{
         authorization_guard::AuthorizationGuard,
-        authorization_guard_with_context::{Create, Read, Update},
+        authorization_guard::{Create, Read, Update},
     },
     user::models::ActiveUser,
 };
