+++ server/domain/src/repository/search_repository.rs
@@ -10,8 +10,7 @@ use crate::{
         models::{ActiveForm, FormLabel},
     },
     search::models::SearchableFieldsWithOperation,
-    types::authorization_guard::AuthorizationGuard,
-    types::authorization_guard_with_context::Read,
+    types::authorization_guard::{AuthorizationGuard, Read},
     user::models::ActiveUser,
 };
