+++ server/domain/src/repository/user_repository.rs
@@ -6,7 +6,7 @@ use uuid::Uuid;
 use crate::{
     types::{
         authorization_guard::AuthorizationGuard,
-        authorization_guard_with_context::{Create, Read, Update},
+        authorization_guard::{Create, Read, Update},
     },
     user::models::{ActiveUser, DiscordUser},
 };
