+++ server/domain/src/types.rs
@@ -1,2 +1,1 @@
 pub mod authorization_guard;
-pub mod authorization_guard_with_context;
