+++ server/domain/src/types/authorization_guard.rs
@@ -1,26 +1,59 @@
 use errors::domain::DomainError;
 
-use crate::{
-    types::authorization_guard_with_context::{
-        Actions, AuthorizationGuardWithContext, AuthorizationGuardWithContextDefinitions, Create,
-        Delete, Read, Update,
-    },
-    user::models::Actor,
-};
+use crate::user::models::Actor;
 
-/// [`Actor`] による `guard_target` に対するアクセスを制御するための定義を提供します。
+pub trait Actions: private::Sealed {}
+
+#[derive(Debug)]
+pub struct Create;
+#[derive(Debug)]
+pub struct Read;
+#[derive(Debug)]
+pub struct Update;
+#[derive(Debug)]
+pub struct Delete;
+
+impl Actions for Create {}
+impl Actions for Read {}
+impl Actions for Update {}
+impl Actions for Delete {}
+
+mod private {
+    pub trait Sealed {}
+
+    impl Sealed for super::Create {}
+    impl Sealed for super::Read {}
+    impl Sealed for super::Update {}
+    impl Sealed for super::Delete {}
+}
+
+/// [`Actor`] による `guard_target` に対するアクセスを制御するための型です。
 ///
-/// [`AuthorizationGuard`] は、Context を必要としない [`AuthorizationGuardWithContext`] であり、
-/// [`AuthorizationGuardWithContext<T, A, ()>`] と同等の機能を提供します。
+/// Action 型パラメータにより、現在許可されている操作の種類を型レベルで表現します。
+// NOTE: Action の変換は以下のようにのみ行うことができます
+//    - Create -> Read
+//    - Create -> Update
+//    - Update <-> Read
+//    - Update または Read -> Delete
+//  これは、データのライフサイクルを考えた時に
+//    - データの新規作成(Create) -> データ読み取り(Read) <-> データ更新(Update) -> データ削除(Delete)
+//  という順序のみ操作が行われるはずであるからです。
+//
+//  仮に Delete から Read へ変換することができるとすると、データの削除操作の実装において
+//  Read 権限を保持しているかつ Delete 権限を持たないユーザーが居る場合に
+//  AuthorizationGuard<T, Delete> から誤って `.into_read()` を呼び出すことで
+//  Read 権限を持つユーザーによってデータが削除されるという事故が発生する可能性があります。
 #[derive(Debug)]
 pub struct AuthorizationGuard<T: AuthorizationGuardDefinitions, A: Actions> {
-    authorization_guard_with_context: AuthorizationGuardWithContext<T, A, ()>,
+    guard_target: T,
+    _phantom_data: std::marker::PhantomData<A>,
 }
 
 impl<T: AuthorizationGuardDefinitions> AuthorizationGuard<T, Create> {
     pub(crate) fn new(guard_target: T) -> Self {
         Self {
-            authorization_guard_with_context: AuthorizationGuardWithContext::new(guard_target),
+            guard_target,
+            _phantom_data: std::marker::PhantomData,
         }
     }
 
@@ -29,8 +62,11 @@ impl<T: AuthorizationGuardDefinitions> AuthorizationGuard<T, Create> {
     where
         F: FnOnce(&'a T) -> R,
     {
-        self.authorization_guard_with_context
-            .try_create(actor, f, &())
+        if self.guard_target.can_create(actor) {
+            Ok(f(&self.guard_target))
+        } else {
+            Err(DomainError::Forbidden)
+        }
     }
 
     /// [`AuthorizationGuardDefinitions::can_create`] の条件で作成操作 `f` を試みます。
@@ -39,20 +75,25 @@ impl<T: AuthorizationGuardDefinitions> AuthorizationGuard<T, Create> {
     where
         F: FnOnce(T) -> R,
     {
-        self.authorization_guard_with_context
-            .try_into_create(actor, f, &())
+        if self.guard_target.can_create(actor) {
+            Ok(f(self.guard_target))
+        } else {
+            Err(DomainError::Forbidden)
+        }
     }
 
     /// [`AuthorizationGuard`] の Action を [`Read`] に変換します。
     pub fn into_read(self) -> AuthorizationGuard<T, Read> {
         AuthorizationGuard {
-            authorization_guard_with_context: self.authorization_guard_with_context.into_read(),
+            guard_target: self.guard_target,
+            _phantom_data: std::marker::PhantomData,
         }
     }
 
     pub fn into_update(self) -> AuthorizationGuard<T, Update> {
         AuthorizationGuard {
-            authorization_guard_with_context: self.authorization_guard_with_context.into_update(),
+            guard_target: self.guard_target,
+            _phantom_data: std::marker::PhantomData,
         }
     }
 }
@@ -63,8 +104,11 @@ impl<T: AuthorizationGuardDefinitions> AuthorizationGuard<T, Update> {
     where
         F: FnOnce(&'a T) -> R,
     {
-        self.authorization_guard_with_context
-            .try_update(actor, f, &())
+        if self.guard_target.can_update(actor) {
+            Ok(f(&self.guard_target))
+        } else {
+            Err(DomainError::Forbidden)
+        }
     }
 
     /// [`AuthorizationGuardDefinitions::can_update`] の条件で更新操作 `f` を試みます。
@@ -73,8 +117,11 @@ impl<T: AuthorizationGuardDefinitions> AuthorizationGuard<T, Update> {
     where
         F: FnOnce(T) -> R,
     {
-        self.authorization_guard_with_context
-            .try_into_update(actor, f, &())
+        if self.guard_target.can_update(actor) {
+            Ok(f(self.guard_target))
+        } else {
+            Err(DomainError::Forbidden)
+        }
     }
 
     /// [`T`] の値に対して map 相当の操作を行います。
@@ -83,21 +130,24 @@ impl<T: AuthorizationGuardDefinitions> AuthorizationGuard<T, Update> {
         F: FnOnce(T) -> T,
     {
         AuthorizationGuard {
-            authorization_guard_with_context: self.authorization_guard_with_context.map(f),
+            guard_target: f(self.guard_target),
+            _phantom_data: std::marker::PhantomData,
         }
     }
 
     /// [`AuthorizationGuard`] の Action を [`Read`] に変換します。
     pub fn into_read(self) -> AuthorizationGuard<T, Read> {
         AuthorizationGuard {
-            authorization_guard_with_context: self.authorization_guard_with_context.into_read(),
+            guard_target: self.guard_target,
+            _phantom_data: std::marker::PhantomData,
         }
     }
 
     /// [`AuthorizationGuard`] の Action を [`Delete`] に変換します。
     pub fn into_delete(self) -> AuthorizationGuard<T, Delete> {
         AuthorizationGuard {
-            authorization_guard_with_context: self.authorization_guard_with_context.into_delete(),
+            guard_target: self.guard_target,
+            _phantom_data: std::marker::PhantomData,
         }
     }
 }
@@ -105,26 +155,35 @@ impl<T: AuthorizationGuardDefinitions> AuthorizationGuard<T, Update> {
 impl<T: AuthorizationGuardDefinitions> AuthorizationGuard<T, Read> {
     /// `actor` が `guard_target` の参照を取得することを試みます。
     pub fn try_read(&self, actor: &Actor) -> Result<&T, DomainError> {
-        self.authorization_guard_with_context.try_read(actor, &())
+        if self.guard_target.can_read(actor) {
+            Ok(&self.guard_target)
+        } else {
+            Err(DomainError::Forbidden)
+        }
     }
 
     /// `actor` が `guard_target` を取得することを試みます。
     pub fn try_into_read(self, actor: &Actor) -> Result<T, DomainError> {
-        self.authorization_guard_with_context
-            .try_into_read(actor, &())
+        if self.guard_target.can_read(actor) {
+            Ok(self.guard_target)
+        } else {
+            Err(DomainError::Forbidden)
+        }
     }
 
     /// [`AuthorizationGuard`] の Action を [`Update`] に変換します。
     pub fn into_update(self) -> AuthorizationGuard<T, Update> {
         AuthorizationGuard {
-            authorization_guard_with_context: self.authorization_guard_with_context.into_update(),
+            guard_target: self.guard_target,
+            _phantom_data: std::marker::PhantomData,
         }
     }
 
     /// [`AuthorizationGuard`] の Action を [`Delete`] に変換します。
     pub fn into_delete(self) -> AuthorizationGuard<T, Delete> {
         AuthorizationGuard {
-            authorization_guard_with_context: self.authorization_guard_with_context.into_delete(),
+            guard_target: self.guard_target,
+            _phantom_data: std::marker::PhantomData,
         }
     }
 }
@@ -135,8 +194,11 @@ impl<T: AuthorizationGuardDefinitions> AuthorizationGuard<T, Delete> {
     where
         F: FnOnce(&'a T) -> R,
     {
-        self.authorization_guard_with_context
-            .try_delete(actor, f, &())
+        if self.guard_target.can_delete(actor) {
+            Ok(f(&self.guard_target))
+        } else {
+            Err(DomainError::Forbidden)
+        }
     }
 
     /// [`AuthorizationGuardDefinitions::can_delete`] の条件で削除操作 `f` を試みます。
@@ -145,8 +207,11 @@ impl<T: AuthorizationGuardDefinitions> AuthorizationGuard<T, Delete> {
     where
         F: FnOnce(T) -> R,
     {
-        self.authorization_guard_with_context
-            .try_into_delete(actor, f, &())
+        if self.guard_target.can_delete(actor) {
+            Ok(f(self.guard_target))
+        } else {
+            Err(DomainError::Forbidden)
+        }
     }
 }
 
@@ -189,27 +254,6 @@ pub trait AuthorizationGuardDefinitions {
     fn can_delete(&self, actor: &Actor) -> bool;
 }
 
-impl<T> AuthorizationGuardWithContextDefinitions<()> for T
-where
-    T: AuthorizationGuardDefinitions,
-{
-    fn can_create(&self, actor: &Actor, _context: &()) -> bool {
-        self.can_create(actor)
-    }
-
-    fn can_read(&self, actor: &Actor, _context: &()) -> bool {
-        self.can_read(actor)
-    }
-
-    fn can_update(&self, actor: &Actor, _context: &()) -> bool {
-        self.can_update(actor)
-    }
-
-    fn can_delete(&self, actor: &Actor, _context: &()) -> bool {
-        self.can_delete(actor)
-    }
-}
-
 impl<T: AuthorizationGuardDefinitions> From<T> for AuthorizationGuard<T, Create> {
     fn from(guard_target: T) -> Self {
         AuthorizationGuard::new(guard_target)
@@ -219,8 +263,8 @@ impl<T: AuthorizationGuardDefinitions> From<T> for AuthorizationGuard<T, Create>
 impl<T: AuthorizationGuardDefinitions> From<T> for AuthorizationGuard<T, Read> {
     fn from(guard_target: T) -> Self {
         Self {
-            authorization_guard_with_context: AuthorizationGuardWithContext::new(guard_target)
-                .into_read(),
+            guard_target,
+            _phantom_data: std::marker::PhantomData,
         }
     }
 }
@@ -228,8 +272,8 @@ impl<T: AuthorizationGuardDefinitions> From<T> for AuthorizationGuard<T, Read> {
 impl<T: AuthorizationGuardDefinitions> From<T> for AuthorizationGuard<T, Update> {
     fn from(guard_target: T) -> Self {
         Self {
-            authorization_guard_with_context: AuthorizationGuardWithContext::new(guard_target)
-                .into_update(),
+            guard_target,
+            _phantom_data: std::marker::PhantomData,
         }
     }
 }
@@ -237,9 +281,8 @@ impl<T: AuthorizationGuardDefinitions> From<T> for AuthorizationGuard<T, Update>
 impl<T: AuthorizationGuardDefinitions> From<T> for AuthorizationGuard<T, Delete> {
     fn from(guard_target: T) -> Self {
         Self {
-            authorization_guard_with_context: AuthorizationGuardWithContext::new(guard_target)
-                .into_read()
-                .into_delete(),
+            guard_target,
+            _phantom_data: std::marker::PhantomData,
         }
     }
 }
