+++ server/infra/resource/src/repository/form_repository_impls/answer_entry_set_repository_impl.rs
@@ -8,7 +8,7 @@ use domain::{
     repository::form::answer_entry_set_repository::AnswerEntrySetRepository,
     types::{
         authorization_guard::AuthorizationGuard,
-        authorization_guard_with_context::{Create, Read, Update},
+        authorization_guard::{Create, Read, Update},
     },
     user::models::Actor,
 };
