+++ server/infra/resource/src/repository/form_repository_impls/form_repository_impl.rs
@@ -7,7 +7,7 @@ use domain::{
     },
     types::{
         authorization_guard::AuthorizationGuard,
-        authorization_guard_with_context::{Create, Read, Update},
+        authorization_guard::{Create, Read, Update},
     },
     user::models::{ActiveUser, Actor},
 };
