+++ server/infra/resource/src/repository/form_repository_impls/message_thread_repository_impl.rs
@@ -4,7 +4,7 @@ use domain::{
     repository::form::message_thread_repository::MessageThreadRepository,
     types::{
         authorization_guard::AuthorizationGuard,
-        authorization_guard_with_context::{Create, Read, Update},
+        authorization_guard::{Create, Read, Update},
     },
     user::models::Actor,
 };
