+++ server/infra/resource/src/repository/notification_repository_impl.rs
@@ -4,7 +4,7 @@ use domain::{
     repository::notification_repository::NotificationRepository,
     types::{
         authorization_guard::AuthorizationGuard,
-        authorization_guard_with_context::{Create, Read, Update},
+        authorization_guard::{Create, Read, Update},
     },
     user::models::{ActiveUser, Actor},
 };
