+++ server/infra/resource/src/repository/user_repository_impl.rs
@@ -3,7 +3,7 @@ use domain::{
     repository::user_repository::UserRepository,
     types::{
         authorization_guard::AuthorizationGuard,
-        authorization_guard_with_context::{Create, Read, Update},
+        authorization_guard::{Create, Read, Update},
     },
     user::models::{ActiveUser, Actor, DiscordUser, DiscordUserId, DiscordUserName},
 };
