+++ server/usecase/src/forms/message.rs
@@ -21,7 +21,7 @@ use domain::{
     },
     types::{
         authorization_guard::AuthorizationGuard,
-        authorization_guard_with_context::{Create, Read, Update},
+        authorization_guard::{Create, Read, Update},
     },
     user::models::{ActiveUser, Actor},
 };
